+++ src/cypress/support/commands.ts
@@ -1,25 +1,34 @@
 import { cy, Cypress } from 'local-cypress'
 import { PrepareForCaptureSettings } from 'src/types';
+import 'cypress-set-device-pixel-ratio';
 
 Cypress.Commands.add('prepareForCapture', (props: PrepareForCaptureSettings) => {
     const { fullUrl, viewport, onPageVisitFunctions, fullPageCapture, options } = props;
 
+    cy.setDevicePixelRatio(options.devicePixelRatio || 1);
     cy.setResolution(viewport);
-    cy.visit(fullUrl);
+    cy.visit(fullUrl, {
+        failOnStatusCode: options.failOnStatusCode ?? true,
+    });
+
     onPageVisitFunctions?.forEach((fn) => fn && fn(cy, Cypress));
     
-    const duration = options.scrollDuration ?? 1000;
+    const scrollSettings = {
+        duration: options.scrollDuration || 1000,
+        ensureScrollable: false,
+    }
 
     if (fullPageCapture) {
-        cy.scrollTo('bottom', { duration, ensureScrollable: false });
-        cy.scrollTo('top', { duration, ensureScrollable: false });
+        cy.scrollTo('bottom', scrollSettings);
+        cy.scrollTo('top', scrollSettings);
         return;
     }     
 
     cy.window().then(win => {
         // We scroll a little even if capture is set to viewport, to trigger any lazy loading/interscetion observer.
-        cy.scrollTo(0, win.innerHeight, { duration: duration / 2, ensureScrollable: false });
-        cy.scrollTo('top', { duration: duration / 2, ensureScrollable: false });
+        cy.scrollTo(0, win.innerHeight, { ...scrollSettings, duration: scrollSettings.duration / 2 });
+        cy.scrollTo('top', { ...scrollSettings, duration: scrollSettings.duration / 2 });
+        
         // Pause at top to allow scrollbar to disappear if it was present.
         cy.wait(100);
     });